pyproject.toml
[project]
name = "imagesegmentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "matplotlib>=3.10.3",
    "onnx>=1.18.0",
    "opendatasets>=0.1.22",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "wandb>=0.20.1",
]
